evaluation/python/requirements.txt
tensorflow
numpy
Pillow
